Implemented popovers using popper.js [ci skip]
