+++ addon/components/bs3/bs-popover/element.js
@@ -1,5 +1,15 @@
 import PopoverElement from 'ember-bootstrap/components/base/bs-popover/element';
+import { computed } from '@ember/object';
 
 export default PopoverElement.extend({
-  classNameBindings: ['placement', 'showHelp:in']
+  popperClassNames: computed('fade', 'placement', 'showHelp', function() {
+    let classes = ['popover', 'ember-bootstrap-popover', this.get('placement')];
+    if (this.get('fade')) {
+      classes.push('fade');
+    }
+    if (this.get('showHelp')) {
+      classes.push('in');
+    }
+    return classes;
+  })
 });
