+++ addon/components/bs4/bs-popover/element.js
@@ -2,11 +2,14 @@ import { computed } from '@ember/object';
 import PopoverElement from 'ember-bootstrap/components/base/bs-popover/element';
 
 export default PopoverElement.extend({
-  classNameBindings: ['placementClass', 'showHelp:show'],
-
-  placementClass: computed('placement', function() {
-    let placement = this.get('placement');
-
-    return `popover-${placement}`;
+  popperClassNames: computed('fade', 'placement', 'showHelp', function() {
+    let classes = ['popover', `popover-${this.get('placement')}`];
+    if (this.get('fade')) {
+      classes.push('fade');
+    }
+    if (this.get('showHelp')) {
+      classes.push('show');
+    }
+    return classes;
   })
 });
