+++ addon/templates/components/common/bs-popover.hbs
@@ -1,15 +1,18 @@
 {{#if inDom}}
-
-  {{#ember-wormhole to="ember-bootstrap-wormhole" renderInPlace=_renderInPlace}}
-
-    {{#bs-popover/element id=overlayId placement=_placement fade=fade showHelp=showHelp title=title class=class}}
-      {{#if hasBlock}}
-        {{yield}}
-      {{else}}
-        {{title}}
-      {{/if}}
-    {{/bs-popover/element}}
-
-  {{/ember-wormhole}}
-
+  {{#bs-popover/element
+    id=overlayId
+    placement=_placement
+    fade=fade
+    showHelp=showHelp
+    title=title
+    class=class
+    renderInPlace=_renderInPlace
+    target=triggerTargetElement
+  }}
+    {{#if hasBlock}}
+      {{yield}}
+    {{else}}
+      {{title}}
+    {{/if}}
+  {{/bs-popover/element}}
 {{/if}}
