+++ addon/templates/components/common/bs-popover/element.hbs
@@ -1,5 +1,14 @@
-<div class="arrow"></div>
-{{#if hasTitle}}
-  <h3 class="popover-title">{{title}}</h3>
-{{/if}}
-<div class="popover-content">{{yield}}</div>
+{{#ember-popper
+  id=id
+  class=popperClass
+  placement=placement
+  renderInPlace=renderInPlace
+  target=target
+  popperContainer="#ember-bootstrap-wormhole"
+}}
+  <div class="arrow"></div>
+  {{#if hasTitle}}
+    <h3 class="popover-title">{{title}}</h3>
+  {{/if}}
+  <div class="popover-content">{{yield}}</div>
+{{/ember-popper}}
